Make rebase_head_name a method, generalize restack
Make Git and CWD injectable in restack::restack,
and make rebase_head_name a method on Git trait
so it can be overridden from tests.
